@@ -5,4 +5,4 @@ py-fsrs
 Py-FSRS is the official Python implementation of the FSRS scheduler algorithm, which can be used to develop spaced repetition systems.
 """
 
-from .fsrs import FSRS, Card, ReviewLog, State, Rating
+from .fsrs import Scheduler, Card, Rating, ReviewLog, State
