@@ -1,9 +1,10 @@
-from fsrs import FSRS, Card, ReviewLog, State, Rating
+from fsrs import Scheduler, Card, ReviewLog, State, Rating
 from datetime import datetime, timedelta, timezone
 import json
 import pytest
+import random
 
-test_w = (
+test_parameters = (
     0.4197,
     1.1869,
     3.0412,
@@ -28,9 +29,7 @@ test_w = (
 
 class TestPyFSRS:
     def test_review_card(self):
-        f = FSRS(w=test_w)
-        card = Card()
-        now = datetime(2022, 11, 29, 12, 30, 0, 0, timezone.utc)
+        scheduler = Scheduler(parameters=test_parameters, enable_fuzzing=False)
 
         ratings = (
             Rating.Good,
@@ -47,15 +46,21 @@ class TestPyFSRS:
             Rating.Good,
             Rating.Good,
         )
+
+        card = Card()
+        review_datetime = datetime(2022, 11, 29, 12, 30, 0, 0, timezone.utc)
+
         ivl_history = []
-
         for rating in ratings:
-            card, _ = f.review_card(card, rating, now)
-            ivl = card.scheduled_days
-            ivl_history.append(ivl)
-            now = card.due
+            card, _ = scheduler.review_card(
+                card=card, rating=rating, review_datetime=review_datetime
+            )
 
-        print(ivl_history)
+            ivl = (card.due - card.last_review).days
+            ivl_history.append(ivl)
+
+            review_datetime = card.due
+
         assert ivl_history == [
             0,
             4,
@@ -73,11 +78,8 @@ class TestPyFSRS:
         ]
 
     def test_memo_state(self):
-        f = FSRS(w=test_w)
-        card = Card()
-        now = datetime(2022, 11, 29, 12, 30, 0, 0, timezone.utc)
+        scheduler = Scheduler(parameters=test_parameters)
 
-        scheduling_cards = f.repeat(card, now)
         ratings = (
             Rating.Again,
             Rating.Good,
@@ -87,34 +89,44 @@ class TestPyFSRS:
             Rating.Good,
         )
         ivl_history = [0, 0, 1, 3, 8, 21]
-        for rating, ivl in zip(ratings, ivl_history):
-            card = scheduling_cards[rating].card
-            now += timedelta(days=ivl)
-            scheduling_cards = f.repeat(card, now)
 
-        assert round(scheduling_cards[Rating.Good].card.stability, 4) == 71.4554
-        assert round(scheduling_cards[Rating.Good].card.difficulty, 4) == 5.0976
+        card = Card()
+        review_datetime = datetime(2022, 11, 29, 12, 30, 0, 0, timezone.utc)
+
+        for rating, ivl in zip(ratings, ivl_history):
+            card, _ = scheduler.review_card(
+                card=card, rating=rating, review_datetime=review_datetime
+            )
+
+            review_datetime += timedelta(days=ivl)
+
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=review_datetime
+        )
+
+        assert round(card.stability) == 72
+        assert round(card.difficulty, 4) == 5.0976
 
     def test_repeat_default_arg(self):
-        f = FSRS()
+        scheduler = Scheduler()
 
-        card_object = Card()
+        card = Card()
 
-        # repeat time is not specified
-        scheduling_cards = f.repeat(card_object)
+        rating = Rating.Good
 
-        card_rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card,
+            rating=rating,
+        )
 
-        card_object = scheduling_cards[card_rating].card
-
-        due = card_object.due
+        due = card.due
 
         time_delta = due - datetime.now(timezone.utc)
 
         assert time_delta.seconds > 500  # due in approx. 8-10 minutes
 
     def test_datetime(self):
-        f = FSRS()
+        scheduler = Scheduler()
         card = Card()
 
         # new cards should be due immediately after creation
@@ -126,11 +138,16 @@ class TestPyFSRS:
 
         # repeating a card with a non-utc, non-timezone-aware datetime object should raise a Value Error
         with pytest.raises(ValueError):
-            f.repeat(card, datetime(2022, 11, 29, 12, 30, 0, 0))
+            scheduler.review_card(
+                card=card,
+                rating=Rating.Good,
+                review_datetime=datetime(2022, 11, 29, 12, 30, 0, 0),
+            )
 
-        # repeat a card with rating good before next tests
-        scheduling_cards = f.repeat(card, datetime.now(timezone.utc))
-        card = scheduling_cards[Rating.Good].card
+        # review a card with rating good before next tests
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
 
         # card object's due and last_review attributes must be timezone aware and UTC
         assert card.due.tzinfo == timezone.utc
@@ -139,7 +156,7 @@ class TestPyFSRS:
         assert card.due >= card.last_review
 
     def test_Card_serialize(self):
-        f = FSRS()
+        scheduler = Scheduler()
 
         # create card object the normal way
         card = Card()
@@ -159,34 +176,32 @@ class TestPyFSRS:
         assert card.to_dict() == copied_card.to_dict()
 
         # (x2) perform the above tests once more with a repeated card
-        scheduling_cards = f.repeat(card, datetime.now(timezone.utc))
-        repeated_card = scheduling_cards[Rating.Good].card
+        reviewed_card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
 
         with pytest.raises(TypeError):
-            json.dumps(repeated_card.__dict__)
+            json.dumps(reviewed_card.__dict__)
 
-        assert type(json.dumps(repeated_card.to_dict())) == str
+        assert type(json.dumps(reviewed_card.to_dict())) == str
 
-        repeated_card_dict = repeated_card.to_dict()
-        copied_repeated_card = Card.from_dict(repeated_card_dict)
+        reviewed_card_dict = reviewed_card.to_dict()
+        copied_reviewed_card = Card.from_dict(reviewed_card_dict)
 
-        assert vars(repeated_card) == vars(copied_repeated_card)
-        assert repeated_card.to_dict() == copied_repeated_card.to_dict()
+        assert vars(reviewed_card) == vars(copied_reviewed_card)
+        assert reviewed_card.to_dict() == copied_reviewed_card.to_dict()
 
         # original card and repeated card are different
-        assert vars(card) != vars(repeated_card)
-        assert card.to_dict() != repeated_card.to_dict()
+        assert vars(card) != vars(reviewed_card)
+        assert card.to_dict() != reviewed_card.to_dict()
 
     def test_ReviewLog_serialize(self):
-        f = FSRS()
+        scheduler = Scheduler()
 
         card = Card()
 
-        # repeat a card to get the review_log
-        scheduling_cards = f.repeat(card)
-        rating = Rating.Again
-        card = scheduling_cards[rating].card
-        review_log = scheduling_cards[rating].review_log
+        # review a card to get the review_log
+        card, review_log = scheduler.review_card(card=card, rating=Rating.Again)
 
         # ReviewLog object is not naturally JSON serializable
         with pytest.raises(TypeError):
@@ -198,14 +213,13 @@ class TestPyFSRS:
         # we can reconstruct a copy of the review_log object equivalent to the original
         review_log_dict = review_log.to_dict()
         copied_review_log = ReviewLog.from_dict(review_log_dict)
-        assert vars(review_log) == vars(copied_review_log)
         assert review_log.to_dict() == copied_review_log.to_dict()
 
-        # (x2) perform the above tests once more with a review_log from a repeated card
-        scheduling_cards = f.repeat(card, datetime.now(timezone.utc))
+        # (x2) perform the above tests once more with a review_log from a reviewed card
         rating = Rating.Good
-        card = scheduling_cards[rating].card
-        next_review_log = scheduling_cards[rating].review_log
+        card, next_review_log = scheduler.review_card(
+            card=card, rating=rating, review_datetime=datetime.now(timezone.utc)
+        )
 
         with pytest.raises(TypeError):
             json.dumps(next_review_log.__dict__)
@@ -215,16 +229,14 @@ class TestPyFSRS:
         next_review_log_dict = next_review_log.to_dict()
         copied_next_review_log = ReviewLog.from_dict(next_review_log_dict)
 
-        assert vars(next_review_log) == vars(copied_next_review_log)
         assert next_review_log.to_dict() == copied_next_review_log.to_dict()
 
         # original review log and next review log are different
-        assert vars(review_log) != vars(next_review_log)
         assert review_log.to_dict() != next_review_log.to_dict()
 
     def test_custom_scheduler_args(self):
-        f = FSRS(
-            w=(
+        scheduler = Scheduler(
+            parameters=(
                 0.4197,
                 1.1869,
                 3.0412,
@@ -245,8 +257,9 @@ class TestPyFSRS:
                 0,
                 0.6468,
             ),
-            request_retention=0.9,
+            desired_retention=0.9,
             maximum_interval=36500,
+            enable_fuzzing=False,
         )
         card = Card()
         now = datetime(2022, 11, 29, 12, 30, 0, 0, timezone.utc)
@@ -269,16 +282,15 @@ class TestPyFSRS:
         ivl_history = []
 
         for rating in ratings:
-            card, _ = f.review_card(card, rating, now)
-            ivl = card.scheduled_days
+            card, _ = scheduler.review_card(card, rating, now)
+            ivl = (card.due - card.last_review).days
             ivl_history.append(ivl)
             now = card.due
 
-        print(ivl_history)
         assert ivl_history == [0, 3, 13, 50, 163, 473, 0, 0, 12, 34, 91, 229, 541]
 
         # initialize another scheduler and verify parameters are properly set
-        w2 = (
+        parameters2 = (
             0.1456,
             0.4186,
             1.1104,
@@ -299,42 +311,400 @@ class TestPyFSRS:
             1.234,
             5.6789,
         )
-        request_retention2 = 0.85
+        desired_retention2 = 0.85
         maximum_interval2 = 3650
-        f2 = FSRS(
-            w=w2,
-            request_retention=request_retention2,
+        scheduler2 = Scheduler(
+            parameters=parameters2,
+            desired_retention=desired_retention2,
             maximum_interval=maximum_interval2,
         )
 
-        assert f2.p.w == w2
-        assert f2.p.request_retention == request_retention2
-        assert f2.p.maximum_interval == maximum_interval2
+        assert scheduler2.parameters == parameters2
+        assert scheduler2.desired_retention == desired_retention2
+        assert scheduler2.maximum_interval == maximum_interval2
 
     def test_retrievability(self):
-        f = FSRS()
+        scheduler = Scheduler()
 
         card = Card()
 
         # retrievabiliy of New card
-        assert card.state == State.New
+        assert card.state == State.Learning
         retrievability = card.get_retrievability()
         assert retrievability == 0
 
         # retrievabiliy of Learning card
-        card, _ = f.review_card(card, Rating.Good)
+        card, _ = scheduler.review_card(card, Rating.Good)
         assert card.state == State.Learning
         retrievability = card.get_retrievability()
         assert 0 <= retrievability <= 1
 
         # retrievabiliy of Review card
-        card, _ = f.review_card(card, Rating.Good)
+        card, _ = scheduler.review_card(card, Rating.Good)
         assert card.state == State.Review
         retrievability = card.get_retrievability()
         assert 0 <= retrievability <= 1
 
         # retrievabiliy of Relearning card
-        card, _ = f.review_card(card, Rating.Again)
+        card, _ = scheduler.review_card(card, Rating.Again)
         assert card.state == State.Relearning
         retrievability = card.get_retrievability()
         assert 0 <= retrievability <= 1
+
+    def test_Scheduler_serialize(self):
+        scheduler = Scheduler()
+
+        # Scheduler objects are json-serializable through its .to_dict() method
+        assert type(json.dumps(scheduler.to_dict())) == str
+
+        # scheduler can be serialized and de-serialized while remaining the same
+        scheduler_dict = scheduler.to_dict()
+        copied_scheduler = Scheduler.from_dict(scheduler_dict)
+        assert vars(scheduler) == vars(copied_scheduler)
+        assert scheduler.to_dict() == copied_scheduler.to_dict()
+
+    def test_good_learning_steps(self):
+        scheduler = Scheduler()
+
+        created_at = datetime.now(timezone.utc)
+        card = Card()
+
+        assert card.state == State.Learning
+        assert card.step == 0
+
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Learning
+        assert card.step == 1
+        assert (
+            round((card.due - created_at).total_seconds() / 100) == 6
+        )  # card is due in approx. 10 minutes (600 seconds)
+
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+        assert card.state == State.Review
+        assert card.step is None
+        assert (
+            round((card.due - created_at).total_seconds() / 3600) >= 24
+        )  # card is due in over a day
+
+    def test_again_learning_steps(self):
+        scheduler = Scheduler()
+
+        created_at = datetime.now(timezone.utc)
+        card = Card()
+
+        assert card.state == State.Learning
+        assert card.step == 0
+
+        rating = Rating.Again
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Learning
+        assert card.step == 0
+        assert (
+            round((card.due - created_at).total_seconds() / 10) == 6
+        )  # card is due in approx. 1 minute (60 seconds)
+
+    def test_hard_learning_steps(self):
+        scheduler = Scheduler()
+
+        created_at = datetime.now(timezone.utc)
+        card = Card()
+
+        assert card.state == State.Learning
+        assert card.step == 0
+
+        rating = Rating.Hard
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Learning
+        assert card.step == 0
+        assert (
+            round((card.due - created_at).total_seconds() / 10) == 33
+        )  # card is due in approx. 5.5 minutes (330 seconds)
+
+    def test_easy_learning_steps(self):
+        scheduler = Scheduler()
+
+        created_at = datetime.now(timezone.utc)
+        card = Card()
+
+        assert card.state == State.Learning
+        assert card.step == 0
+
+        rating = Rating.Easy
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Review
+        assert card.step is None
+        assert (
+            round((card.due - created_at).total_seconds() / 86400) >= 1
+        )  # card is due in at least 1 full day
+
+    def test_review_state(self):
+        scheduler = Scheduler(enable_fuzzing=False)
+
+        card = Card()
+
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Review
+        assert card.step is None
+
+        prev_due = card.due
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Review
+        assert (
+            round((card.due - prev_due).total_seconds() / 3600) >= 24
+        )  # card is due in at least 1 full day
+
+        # rate the card again
+        prev_due = card.due
+        rating = Rating.Again
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Relearning
+        assert (
+            round((card.due - prev_due).total_seconds() / 60) == 10
+        )  # card is due in 10 minutes
+
+    def test_relearning(self):
+        scheduler = Scheduler(enable_fuzzing=False)
+
+        card = Card()
+
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        prev_due = card.due
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        # rate the card again
+        prev_due = card.due
+        rating = Rating.Again
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Relearning
+        assert card.step == 0
+        assert (
+            round((card.due - prev_due).total_seconds() / 60) == 10
+        )  # card is due in 10 minutes
+
+        prev_due = card.due
+        rating = Rating.Again
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Relearning
+        assert card.step == 0
+        assert (
+            round((card.due - prev_due).total_seconds() / 60) == 10
+        )  # card is due in 10 minutes
+
+        prev_due = card.due
+        rating = Rating.Good
+        card, _ = scheduler.review_card(
+            card=card, rating=rating, review_datetime=card.due
+        )
+
+        assert card.state == State.Review
+        assert card.step is None
+        assert (
+            round((card.due - prev_due).total_seconds() / 3600) >= 24
+        )  # card is due in at least 1 full day
+
+    def test_fuzz(self):
+        scheduler = Scheduler()
+
+        # seed 1
+        random.seed(42)
+
+        card = Card()
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=card.due
+        )
+        prev_due = card.due
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=card.due
+        )
+        interval = card.due - prev_due
+
+        assert interval.days == 16
+
+        # seed 2
+        random.seed(12345)
+
+        card = Card()
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=card.due
+        )
+        prev_due = card.due
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=card.due
+        )
+        interval = card.due - prev_due
+
+        assert interval.days == 15
+
+    def test_no_learning_steps(self):
+        scheduler = Scheduler(learning_steps=())
+
+        assert len(scheduler.learning_steps) == 0
+
+        card = Card()
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Again, review_datetime=datetime.now(timezone.utc)
+        )
+
+        assert card.state == State.Review
+        interval = (card.due - card.last_review).days
+        assert interval >= 1
+
+    def test_no_relearning_steps(self):
+        scheduler = Scheduler(relearning_steps=())
+
+        assert len(scheduler.relearning_steps) == 0
+
+        card = Card()
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
+        assert card.state == State.Learning
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=card.due
+        )
+        assert card.state == State.Review
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Again, review_datetime=card.due
+        )
+        assert card.state == State.Review
+
+        interval = (card.due - card.last_review).days
+        assert interval >= 1
+
+    def test_one_card_multiple_schedulers(self):
+        scheduler_with_two_learning_steps = Scheduler(
+            learning_steps=(timedelta(minutes=1), timedelta(minutes=10))
+        )
+        scheduler_with_no_learning_steps = Scheduler(learning_steps=())
+
+        card = Card()
+
+        assert len(scheduler_with_two_learning_steps.learning_steps) == 2
+        card, _ = scheduler_with_two_learning_steps.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
+        assert card.state == State.Learning
+        assert card.step == 1
+
+        assert len(scheduler_with_no_learning_steps.learning_steps) == 0
+        card, _ = scheduler_with_no_learning_steps.review_card(
+            card=card, rating=Rating.Again, review_datetime=datetime.now(timezone.utc)
+        )
+        assert card.state == State.Review
+        assert card.step is None
+
+        scheduler_with_two_relearning_steps = Scheduler(
+            relearning_steps=(
+                timedelta(minutes=1),
+                timedelta(minutes=10),
+                timedelta(minutes=15),
+            )
+        )
+        scheduler_with_no_relearning_steps = Scheduler(relearning_steps=())
+
+        assert len(scheduler_with_two_relearning_steps.relearning_steps) == 3
+        card, _ = scheduler_with_two_relearning_steps.review_card(
+            card=card, rating=Rating.Again, review_datetime=datetime.now(timezone.utc)
+        )
+        assert card.state == State.Relearning
+        assert card.step == 0
+
+        card, _ = scheduler_with_two_relearning_steps.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
+        assert card.state == State.Relearning
+        assert card.step == 1
+
+        card, _ = scheduler_with_two_relearning_steps.review_card(
+            card=card, rating=Rating.Good, review_datetime=datetime.now(timezone.utc)
+        )
+        assert card.state == State.Relearning
+        assert card.step == 2
+
+        card, _ = scheduler_with_no_relearning_steps.review_card(
+            card=card, rating=Rating.Again, review_datetime=datetime.now(timezone.utc)
+        )
+        assert card.state == State.Review
+        assert card.step is None
+
+    def test_maximum_interval(self):
+        maximum_interval = 100
+        scheduler = Scheduler(maximum_interval=maximum_interval)
+
+        card = Card()
+
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Easy, review_datetime=card.due
+        )
+        assert (card.due - card.last_review).days <= scheduler.maximum_interval
+
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=card.due
+        )
+        assert (card.due - card.last_review).days <= scheduler.maximum_interval
+
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Easy, review_datetime=card.due
+        )
+        assert (card.due - card.last_review).days <= scheduler.maximum_interval
+
+        card, _ = scheduler.review_card(
+            card=card, rating=Rating.Good, review_datetime=card.due
+        )
+        assert (card.due - card.last_review).days <= scheduler.maximum_interval
